Move makeUserEntity to interactions file
