+++ src/main/java/com/google/sps/Utils.java
@@ -14,8 +14,6 @@
 
 package com.google.sps;
 
-import com.google.appengine.api.datastore.DatastoreService;
-import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.GeoPt;
 import com.google.gson.Gson;
@@ -116,6 +114,13 @@ public class Utils {
     return new GeoPt((float) location.lat, (float) location.lng);
   }
 
+  public static GeoPt getGeopt(String location) {
+    if (location == null) {
+      return null;
+    }
+    return getGeopt(getLatLng(location));
+  }
+
   /**
    * Calculates the distance in km between two locations using driving routes generated by the
    * Google Maps API.
@@ -161,34 +166,6 @@ public class Utils {
       return -1;
     }
     return getDistance(start, end);
-  }
-
-  /**
-   * Creates and returns a new Entity for a given userId.
-   *
-   * @param userId to identify this user in datastore
-   * @param location user's location
-   * @param addToDatastore if true, will add this entity to datastore as well
-   */
-  public static Entity makeUserEntity(String userId, String location, boolean addToDatastore) {
-    Entity userEntity = new Entity("User", userId);
-    userEntity.setProperty("firebaseID", userId);
-    userEntity.setProperty("location", location);
-    if (addToDatastore) {
-      DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-      datastore.put(userEntity);
-    }
-    return userEntity;
-  }
-
-  /**
-   * Creates and returns a new Entity for a given userId, no location given.
-   *
-   * @param userId to identify this user in datastore
-   * @param addToDatastore if true, will add this entity to datastore as well
-   */
-  public static Entity makeUserEntity(String userId, boolean addToDatastore) {
-    return makeUserEntity(userId, "", addToDatastore);
   }
 
   /** Orders a map from greatest to least based off its values. */
